Add files via upload
Designed  deep learning models with Python’s Keras library that give an accuracy between 78 and 83%. The models were modified by changing the number of hidden layers and by changing the number of nodes in each layer. The validation_accuracy and training accuracy were compared with respect to number of epochs.